use docker-machine for ssh commands
docker-machine already has built-in support for issuing ssh commands to
the vm. switching to docker-machine invocations helps the connection to
the vm to be a little less brittle.

unfortunately, fsevents_to_vm still uses the ssh config to connect, so
this cannot be completely removed yet.

additionally, the docker-machine inspect format was changed in 0.5.0,
which changed how the store path needed to be retrieved.
