+++ cli/dinghy/machine.rb
@@ -2,6 +2,14 @@ require 'dinghy/constants'
 require 'json'
 
 class Machine
+  class CommandFailed < RuntimeError
+    attr_reader :exitstatus
+    def initialize(message, exitstatus)
+      super(message)
+      @exitstatus = exitstatus
+    end
+  end
+
   def create(options = {})
     provider = options['provider']
 
@@ -27,7 +35,7 @@ class Machine
       raise("There was an error bringing up the VM. Dinghy cannot continue.")
     end
 
-    Ssh.new(self).write_ssh_config!
+    write_ssh_config!
   end
 
   def host_ip
@@ -43,7 +51,12 @@ class Machine
   end
 
   def store_path
-    inspect['StorePath']
+    driver = inspect['Driver']
+    if driver.key?('StorePath')
+      File.join(driver['StorePath'], 'machines', driver['MachineName'])
+    else
+      inspect['StorePath']
+    end
   end
 
   def inspect
@@ -73,7 +86,33 @@ class Machine
   end
 
   def ssh(*command)
-    Ssh.new(self).run(*command)
+    if command.empty?
+      exec('ssh', machine_name)
+    else
+      system('ssh', machine_name, '--', *command)
+      fail CommandFailed.new("Error executing command: #{command}", $?.exitstatus) unless $?.success?
+    end
+  end
+
+  def write_ssh_config!
+    File.open(ssh_config_path, 'wb') { |f| f.write(ssh_config) }
+  end
+
+  def ssh_config
+    # Constructing the IdentityFile path ourselves is a recipe for future
+    # sadness, but I haven't found a way to get it out of docker-machine.
+    <<-SSH
+Host dinghy
+  HostName #{vm_ip}
+  User docker
+  Port 22
+  UserKnownHostsFile /dev/null
+  StrictHostKeyChecking no
+  PasswordAuthentication no
+  IdentityFile #{store_path}/id_rsa
+  IdentitiesOnly yes
+  LogLevel ERROR
+    SSH
   end
 
   def halt
@@ -97,6 +136,10 @@ class Machine
     Kernel.system("docker-machine", *cmd)
   end
 
+  def exec(*cmd)
+    Kernel.exec('docker-machine', *cmd)
+  end
+
   def machine_name
     'dinghy'
   end
@@ -114,6 +157,11 @@ class Machine
   end
 
   protected
+
+  def ssh_config_path
+    # this is hard-coded inside the fsevents_to_vm plist, as well
+    HOME_DINGHY+"ssh-config"
+  end
 
   def configure_new_machine(provider)
     if provider == "virtualbox"
